requirements.txt
django==5.0
psycopg2==2.9.9
python-decouple==3.8
Pillow==10.1.0
djangorestframework==3.14.0
django-cleanup==8.0.0
django-rest-authtoken==2.1.4